[fix] Replace getRandomUnitVector() with getRandomVectorInUnitSphere()
The distribution of these two functions is quite different which causes
the generated images have quite different effect. One is just a random
point, while the other one is a random point that is located within an
unit sphere.

Changes to main.cpp:

- Replace getRandomUnitVector() with getRandomVectorInUnitSphere()

Changes to module Utility:

- Add a new function - getRandomDoubleInRange

Changes to class Vector3:

- Remove unused static method getRandomUnitVector()
- Fix the comment of getRandomVector(): [0, 1] -> [0, 1)
- Add a new static method getRandomVectorInUnitSphere()
- Add a new method squaredLength() for determing whether the squared
  length of a random vector is unit or not
    - Add the related unittest