[feat] Take fuzziness into account in class Metal
Metal can reflects almost all wavelengths of visible ray. However, the
metal may not always be perfectly smooth. And since the fuzziness causes
the reflected ray to scatter, we need to add a new member to represent
the fuzziness (smoothness) of the metal.

Changes to main.cpp:

- Add values of fuzziness in constructors of Metal object

Changes to class Metal:

- Add a new member - m_fuzziness
- Apply the effect of fuzziness in method getScatteredRay